+++ signing/src/main/java/tech/pegasys/web3signer/signing/config/metadata/HashicorpSigningMetadata.java
@@ -12,6 +12,7 @@
  */
 package tech.pegasys.web3signer.signing.config.metadata;
 
+import tech.pegasys.web3signer.keystorage.hashicorp.VaultAuthMethod;
 import tech.pegasys.web3signer.signing.ArtifactSigner;
 import tech.pegasys.web3signer.signing.KeyType;
 
@@ -37,11 +38,16 @@ public class HashicorpSigningMetadata extends SigningMetadata {
   private Path tlsKnownServerFile = null;
   private HttpClient.Version httpProtocolVersion;
 
+  private VaultAuthMethod authMethod = VaultAuthMethod.TOKEN;
+  private String kubernetesRole;
+  private Path kubernetesServiceAccountTokenPath;
+  private String kubernetesAuthPath;
+
   @JsonCreator
   public HashicorpSigningMetadata(
       @JsonProperty(value = "serverHost", required = true) final String serverHost,
       @JsonProperty(value = "keyPath", required = true) final String keyPath,
-      @JsonProperty(value = "token", required = true) final String token,
+      @JsonProperty(value = "token") final String token,
       @JsonProperty(value = "keyType") final KeyType keyType) {
     super(TYPE, keyType != null ? keyType : KeyType.BLS);
     this.serverHost = serverHost;
@@ -79,6 +85,28 @@ public class HashicorpSigningMetadata extends SigningMetadata {
     this.httpProtocolVersion = httpProtocolVersion;
   }
 
+  @JsonSetter("authMethod")
+  public void setAuthMethod(final VaultAuthMethod authMethod) {
+    if (authMethod != null) {
+      this.authMethod = authMethod;
+    }
+  }
+
+  @JsonSetter("kubernetesRole")
+  public void setKubernetesRole(final String kubernetesRole) {
+    this.kubernetesRole = kubernetesRole;
+  }
+
+  @JsonSetter("kubernetesServiceAccountTokenPath")
+  public void setKubernetesServiceAccountTokenPath(final Path kubernetesServiceAccountTokenPath) {
+    this.kubernetesServiceAccountTokenPath = kubernetesServiceAccountTokenPath;
+  }
+
+  @JsonSetter("kubernetesAuthPath")
+  public void setKubernetesAuthPath(final String kubernetesAuthPath) {
+    this.kubernetesAuthPath = kubernetesAuthPath;
+  }
+
   public String getServerHost() {
     return serverHost;
   }
@@ -113,6 +141,22 @@ public class HashicorpSigningMetadata extends SigningMetadata {
 
   public HttpClient.Version getHttpProtocolVersion() {
     return httpProtocolVersion;
+  }
+
+  public VaultAuthMethod getAuthMethod() {
+    return authMethod;
+  }
+
+  public String getKubernetesRole() {
+    return kubernetesRole;
+  }
+
+  public Path getKubernetesServiceAccountTokenPath() {
+    return kubernetesServiceAccountTokenPath;
+  }
+
+  public String getKubernetesAuthPath() {
+    return kubernetesAuthPath;
   }
 
   @Override